requirements.txt
python-dotenv
streamlit
zhipuai